# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: number;
  role: MessageRole;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface ClearResult {
  deleted: number;
}

export interface ChatStats {
  totalMessages: number;
  userMessages: number;
  wifeMessages: number;
  /** @nullable */
  firstMessageAt: string | null;
  /** @nullable */
  lastMessageAt: string | null;
}