# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: getMessages
      tags: [chat]
      summary: Get conversation history
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"

    post:
      operationId: sendMessage
      tags: [chat]
      summary: Send a message and get AI response (SSE streaming)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "200":
          description: SSE stream of response chunks
          content:
            text/event-stream:
              schema:
                type: string
  /chat/messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [chat]
      summary: Delete a message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
      required:
        - id
        - role
        - content
        - createdAt
    MessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
      required:
        - content
...